First Commit! Changed the loading screen.
First commit! However this is far from the first iteration of the project. In fact, before this project I knew next to nothing about web development, javascript, html, webvr, or really any of the technology I ended up using. I've learned a ton working on this, and I hope it shows!