Removed swapping Layers Logic
I removed old logic for swapping layers. I am abandoning the layer-based approach. It is overly complicated and unnecessary.  I also no longer create the icon toggle button in the layer script. It exists in our a-scene. This results in better readability and gets us closer the single responsibility principle.